Operations & Settings, Mobile fixes
