+++ src/components/settings/Tokens.scss
@@ -8,14 +8,17 @@
 
   .panel-head {
     display: flex;
+    flex-wrap: wrap;
     flex-direction: row;
+    align-items: center;
     color: $color-primary-0;
+    padding: 24px 16px;
 
     .head-title {
       display: block;
       flex: 1 1 auto;
       margin: 0;
-      padding: 24px 16px 0px 16px;
+      padding: 0;
       font-size: 24px;
       line-height: 28px;
       font-weight: normal;
@@ -23,7 +26,6 @@
 
     .head-actions {
       flex: 0 0 auto;
-      padding: 18px 16px 0px 16px;
     }
   }
 
@@ -67,6 +69,40 @@
         border-top: 1px solid $color-hr-1;
         padding: 16px 16px;
 
+        .cell-title {
+          display: none;
+          vertical-align: baseline;
+          opacity: 0.6;
+        }
+
+        .cell-name {
+          display: inline-flex;
+          vertical-align: middle;
+          position: relative;
+          flex-wrap: nowrap;
+          align-items: center;
+
+          .name-icon {
+            position: relative;
+            display: 0 0 auto;
+            background: $color-panel-bg-0;
+            width: 36px;
+            height: 24px;
+            margin-right: 12px;
+            border-radius: 2px;
+
+            .icon-content {
+              @include position(absolute, 0, 0, 0, 0);
+              background: transparent none no-repeat center center;
+              background-size: contain;
+            }
+          }
+
+          .name-title {
+            flex: 1 1 auto;
+          }
+        }
+
         &.table-cell-name {
         }
         &.table-cell-address {
@@ -98,6 +134,45 @@
       white-space: nowrap;
       overflow: hidden;
       text-overflow: ellipsis;
+    }
+  }
+
+  @include sm-only {
+    .panel-table {
+      display: block;
+
+      .table-head {
+        display: none;
+      }
+
+      .table-body {
+        display: block;
+
+        .table-row {
+          display: block;
+          border-top: 1px solid $color-hr-1;
+
+          .table-cell {
+            display: block;
+            border-top: none;
+            padding: 16px 16px 0px 16px;
+
+            .cell-title {
+              display: inline-block;
+            }
+
+            &.table-cell-actions {
+              .actions {
+                display: flex;
+                flex-wrap: nowrap;
+              }
+            }
+            &:last-child {
+              padding-bottom: 16px;
+            }
+          }
+        }
+      }
     }
   }
 }
